production build

+++ src/products/products.service.ts
@@ -1,10 +1,8 @@
-import { Injectable, NotFoundException } from "@nestjs/common";
+import { Injectable, NotFoundException, UnauthorizedException } from "@nestjs/common";
 import { InjectModel } from "@nestjs/mongoose";
 import { ProductDto } from "src/products/dto/product.dto";
 import { Product } from "./schemas/product.schema";
-import mongoose from "mongoose";
-import { User } from "src/auth/schemas/user.schema";
-
+import mongoose, { Types } from "mongoose";
 
 @Injectable()
 export class ProductService{
@@ -12,9 +10,13 @@ export class ProductService{
         @InjectModel(Product.name)
         private products: mongoose.Model<Product>
     ) {}
-    async create(product: ProductDto): Promise<Product> {
-        const res = await this.products.create(product);
-        return res;
+    async create(productDto: ProductDto, userId: string): Promise<Product> {
+        const userObjectId = new Types.ObjectId(userId);
+        const product = new this.products({
+            ...productDto,
+            author: userObjectId
+        });
+        return product.save();
     }
     async findAll(): Promise<Product[]>{
         const product = await this.products.find();
@@ -34,7 +36,14 @@ export class ProductService{
 
         return product
     }
-    async delete(id: string): Promise<Product>{
+    async delete(id: string, userId: string){
+        const product = await this.products.findById(id);
+        const userObjectId = new Types.ObjectId(userId);
+
+        if(!product.author.equals(userObjectId)){
+            throw new UnauthorizedException('You are not authorized to delete products that you did not make!')
+        }
+        
         return await this.products.findByIdAndDelete(id);
     };
 };
